backend/requirements.txt

# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2

# Text analysis
textstat==0.7.3
nltk==3.8.1

# Utilities
python-dotenv==1.0.0
pathlib==1.0.1

# Development and testing
pytest==7.4.0
pytest-cov==4.1.0
